+++ src/app/podcasts/[id]/page.tsx
@@ -9,7 +9,7 @@ import {
   formatMillisecondsToMinutesAndSeconds,
   DateFormats,
 } from "@utils/static/formatDate"
-import { PodcastEpisode } from "types/interfaces/interface"
+import { PodcastEpisode } from "types/interfaces/PodcastEpisode"
 import { SpotifyService } from "@/services/SpotifyService"
 
 const EpisodePage = () => {
